+++ R/Mmatrix.R
@@ -1,7 +1,7 @@
 #' Calculate \eqn{\mathbf{M}} matrix
 #'
 #' @description
-#' Calculate \eqn{\mathbf{M}} matrix for stauted CDMs (de la Torre, 2011). The \eqn{\mathbf{M}} matrix is a matrix used to 
+#' Calculate \eqn{\mathbf{M}} matrix for saturated CDMs (de la Torre, 2011). The \eqn{\mathbf{M}} matrix is a matrix used to 
 #' represent the interaction mechanisms between attributes.  
 #'
 #' @param K The number of attributes. Can be \code{NULL} if the argument \code{pattern} is not \code{NULL}.
@@ -22,7 +22,9 @@
 #'
 #' library(Qval)
 #'
-#' example.Mmatrix <-  get.Mmatrix(K = 3)
+#' Mmatrix <-  get.Mmatrix(K = 3)
+#' 
+#' print(Mmatrix)
 #'
 #' @export
 #' @importFrom GDINA attributepattern