+++ R/sim.data.R
@@ -1,4 +1,4 @@
-#' generate response
+#' Generate Response
 #'
 #' @description
 #' randomly generate response matrix according to certain conditions,
@@ -70,9 +70,9 @@
 #'   P1 = runif(I, 0.8, 1.0)
 #' )
 #'
-#' data <- sim.data(Q = Q, N = 10, IQ=IQ, model = "GDINA", distribute = "uniform")
+#' data.obj <- sim.data(Q = Q, N = 100, IQ=IQ, model = "GDINA", distribute = "uniform")
 #'
-#' print(data$dat)
+#' print(data.obj$dat)
 #'
 #'################################################################
 #'#                           Example 2                          #
@@ -88,11 +88,11 @@
 #'   P1 = runif(I, 0.8, 1.0)
 #' )
 #'
-#' example_cutoffs <- sample(qnorm(c(1:K)/(K+1)), ncol(Q))
-#' data <- sim.data(Q = Q, N = 10, IQ=IQ, model = "GDINA", distribute = "mvnorm",
-#'                  control = list(sigma = 0.5, cutoffs = example_cutoffs))
+#' cutoffs <- sample(qnorm(c(1:K)/(K+1)), ncol(Q))
+#' data.obj <- sim.data(Q = Q, N = 10, IQ=IQ, model = "GDINA", distribute = "mvnorm",
+#'                  control = list(sigma = 0.5, cutoffs = cutoffs))
 #'
-#' print(data$dat)
+#' print(data.obj$dat)
 #'
 #'#################################################################
 #'#                            Example 3                          #
@@ -108,12 +108,12 @@
 #'   P1 = runif(I, 0.8, 1.0)
 #' )
 #'
-#' example_theta <- rnorm(10, 0, 1)
-#' example_b <- seq(-1.5,1.5,length.out=K)
-#' data <- sim.data(Q = Q, N = 10, IQ=IQ, model = "GDINA", distribute = "horder",
-#'                  control = list(theta = example_theta, a = 1.5, b = example_b))
+#' theta <- rnorm(10, 0, 1)
+#' b <- seq(-1.5,1.5,length.out=K)
+#' data.obj <- sim.data(Q = Q, N = 10, IQ=IQ, model = "GDINA", distribute = "horder",
+#'                  control = list(theta = theta, a = 1.5, b = b))
 #'
-#' print(data$dat)
+#' print(data.obj$dat)
 #'
 #' @export
 #' @importFrom GDINA attributepattern
@@ -217,7 +217,16 @@ sim.data <- function(Q=NULL, N=NULL, IQ=list(P0=NULL, P1=NULL),
               higher.order.parm=higher.order.parm, 
               mvnorm.parm=mvnorm.parm, 
               LCprob.parm=LCprob.parm, 
-              call = simCall)
+              call = simCall, 
+              arguments = list(
+                Q=Q, 
+                N=N, 
+                IQ=IQ,
+                model=model, 
+                distribute=distribute, 
+                control = control,
+                verbose = verbose
+              ))
    
    class(out) <- "sim.data"